fix: document banner
