+++ frontend/appflowy_tauri/src/appflowy_app/components/document/DocumentBanner/DocumentIcon.tsx
