+++ frontend/appflowy_tauri/src/appflowy_app/components/document/DocumentBanner/cover/config.ts
