+++ frontend/appflowy_tauri/src/appflowy_app/components/document/DocumentTitle/index.tsx
@@ -2,11 +2,10 @@ import React, { useState } from 'react';
 import { useDocumentTitle } from './DocumentTitle.hooks';
 import TextBlock from '../TextBlock';
 import { useTranslation } from 'react-i18next';
-import TitleButtonGroup from './TitleButtonGroup';
-import DocumentTopPanel from './DocumentTopPanel';
+import DocumentBanner from '$app/components/document/DocumentBanner';
 
 export default function DocumentTitle({ id }: { id: string }) {
-  const { node, onUpdateCover, onUpdateIcon } = useDocumentTitle(id);
+  const { node } = useDocumentTitle(id);
   const { t } = useTranslation();
   const [hover, setHover] = useState(false);
 
@@ -14,14 +13,7 @@ export default function DocumentTitle({ id }: { id: string }) {
 
   return (
     <div className={'flex flex-col'} onMouseEnter={() => setHover(true)} onMouseLeave={() => setHover(false)}>
-      <DocumentTopPanel onUpdateCover={onUpdateCover} onUpdateIcon={onUpdateIcon} node={node} />
-      <div
-        style={{
-          opacity: hover ? 1 : 0,
-        }}
-      >
-        <TitleButtonGroup node={node} onUpdateCover={onUpdateCover} onUpdateIcon={onUpdateIcon} />
-      </div>
+      <DocumentBanner id={node.id} hover={hover} />
       <div data-block-id={node.id} className='doc-title relative text-4xl font-bold'>
         <TextBlock placeholder={t('document.title.placeholder')} childIds={[]} node={node} />
       </div>
